moved weather functions into separate folder
