+++ src/campsite-detail.js
@@ -1,8 +1,8 @@
 import { loadHeader } from './header-component.js';
 import makeDetailTemplate, { makeImageTemplate } from './detail-component.js';
 import { makeFacilityUrl, makeMediaUrl } from '../src/make-detail-url.js';
-import makeForecastCard from './weather/make-forecast-card.js';
-// import data from '../data/sample-forecast-data.js';
+import loadWeather from './weather/make-forecast-card.js';
+import makeWeatherUrl from './weather/make-weather-url.js';
 
 const searchParams = new URLSearchParams(window.location.search);
 const facilityID = searchParams.get('facilityId');
@@ -11,13 +11,6 @@ const lon = searchParams.get('lon');
 const facilityUrl = makeFacilityUrl(facilityID);
 const mediaUrl = makeMediaUrl(facilityID);
 const weatherUrl = makeWeatherUrl(lat, lon);
-
-
-function makeWeatherUrl(lat, lon) {
-    const path = `https://api.openweathermap.org/data/2.5/forecast?lat=${lat}&lon=${lon}&APPID=02fd61fe81d6deddb87cd6c131a9148d&units=imperial`;
-    const url = new URL(path);
-    return url.toString();
-}
 
 const user = {
     displayName: 'Anna',
@@ -59,11 +52,5 @@ function loadImages(data) {
     main.appendChild(dom);
 }
 
-function loadWeather(forecast) {
-    const forecastDisplay = document.getElementById('forecast-list');
-    forecast.list.forEach(day => {
-        const dom = makeForecastCard(day);
-        forecastDisplay.appendChild(dom);
-    });
-}
+
